+++ packages/yourloops/components/icons/RoundedHospitalIcon.tsx
@@ -1,6 +1,5 @@
-/**
+/*
  * Copyright (c) 2022, Diabeloop
- * Rounded Hospital Icon
  *
  * All rights reserved.
  *