+++ packages/yourloops/lib/config.ts
@@ -1,6 +1,5 @@
-/**
- * Copyright (c) 2021, Diabeloop
- * Yourloops configuration
+/*
+ * Copyright (c) 2021-2022, Diabeloop
  *
  * All rights reserved.
  *