+++ packages/yourloops/test/unit/lib/metrics.test.tsx
@@ -1,6 +1,5 @@
-/**
- * Copyright (c) 2021, Diabeloop
- * metrics tests
+/*
+ * Copyright (c) 2021-2022, Diabeloop
  *
  * All rights reserved.
  *