+++ packages/yourloops/components/patient/models.ts
@@ -1,6 +1,5 @@
-/**
- * Copyright (c) 2021, Diabeloop
- * Models for patients list
+/*
+ * Copyright (c) 2021-2022, Diabeloop
  *
  * All rights reserved.
  *