+++ packages/yourloops/lib/data/models.ts
@@ -1,6 +1,5 @@
-/**
- * Copyright (c) 2021, Diabeloop
- * Data API Models
+/*
+ * Copyright (c) 2021-2022, Diabeloop
  *
  * All rights reserved.
  *